--- modules/cross-ranker-transformers/additional/models/models.go
@@ -1,18 +1,0 @@
-//                           _       _
-// __      _____  __ ___   ___  __ _| |_ ___
-// \ \ /\ / / _ \/ _` \ \ / / |/ _` | __/ _ \
-//  \ V  V /  __/ (_| |\ V /| | (_| | ||  __/
-//   \_/\_/ \___|\__,_| \_/ |_|\__,_|\__\___|
-//
-//  Copyright © 2016 - 2023 Weaviate B.V. All rights reserved.
-//
-//  CONTACT: hello@weaviate.io
-//
-
-package models
-
-// Answer used in qna module to represent
-// the answer to a given question
-type RankResult struct {
-	Score *float64 `json:"score,omitempty"`
-}
